requirements2.txt
amqp==1.4.5
anyjson==0.3.3
eventlet==0.17.4
Flask==0.10.1
Flask-Admin==1.0.8
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.2
greenlet==0.4.6
itsdangerous==0.24
Jinja2==2.7.3
kombu==3.0.21
MarkupSafe==0.23
MySQL-python==1.2.5
pycrypto==2.6.1
SQLAlchemy==0.9.6
Werkzeug==0.9.6
WTForms==2.0.1
zope.interface==4.1.1
